EasyLearn/serializers.py
from rest_framework import serializers
from EasyLearn.models import Post
from .models import Category

class PostSerializer(serializers.ModelSerializer):

    class Meta:
        model = Post
        fields = '__all__'
    
class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ('id', 'name', 'description')