requirements.txt
django
djangoretframework
djoser
djangorestframework_simplejwt
social-auth-app-django